feat(play): Refactoring and spotify playlist support
